current issues see descript todo.jsx
modal only partially opens on long press unless numerous entries are made. also, only passes back correctly the taskID that is selected and does take it on return from the modal as well. not sure if it is not passing all information or if modal is just not accepting all information